requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements.txt --output-file requirements.txt --no-hashes
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.2
    # via
    #   aiohttp-retry
    #   twilio
aiohttp-retry==2.9.1
    # via twilio
aiosignal==1.4.0
    # via aiohttp
annotated-doc==0.0.3
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   google-genai
    #   httpx
    #   starlette
attrs==25.4.0
    # via aiohttp
cachetools==6.2.1
    # via google-auth
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via uvicorn
colorama==0.4.6 ; sys_platform == 'win32'
    # via click
fastapi==0.120.1
    # via fiverr-brianley
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
google-api-core==2.28.1
    # via google-cloud-dialogflow-cx
google-auth==2.42.0
    # via
    #   google-api-core
    #   google-cloud-dialogflow-cx
    #   google-genai
google-cloud-dialogflow-cx==2.0.0
    # via fiverr-brianley
google-genai==1.46.0
    # via fiverr-brianley
googleapis-common-protos==1.71.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.76.0
    # via
    #   google-api-core
    #   google-cloud-dialogflow-cx
    #   grpcio-status
grpcio-status==1.76.0
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   fiverr-brianley
    #   google-genai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-dialogflow-cx
protobuf==6.33.0
    # via
    #   google-api-core
    #   google-cloud-dialogflow-cx
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.12.3
    # via
    #   fastapi
    #   google-genai
pydantic-core==2.41.4
    # via pydantic
pyjwt==2.10.1
    # via twilio
python-dotenv==1.2.1
    # via fiverr-brianley
requests==2.32.5
    # via
    #   google-api-core
    #   google-genai
    #   twilio
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via anyio
starlette==0.49.1
    # via fastapi
tenacity==9.1.2
    # via google-genai
twilio==9.8.5
    # via fiverr-brianley
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   fastapi
    #   google-genai
    #   grpcio
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.5.0
    # via requests
uvicorn==0.38.0
    # via fiverr-brianley
websockets==15.0.1
    # via google-genai
yarl==1.22.0
    # via aiohttp
